get endpoints works now id is long added run.sh script for local running

+++ src/main/kotlin/db/model/Movie.kt
@@ -3,4 +3,4 @@ package zero.network.db.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Movie(val id: Int? = null, val name: String, val year: Int)
+data class Movie(val id: Long? = null, val name: String, val year: Int)
